move: firebase as shared lib
