+++ apps/web/src/app/index.tsx
@@ -4,7 +4,7 @@ import React, { Suspense } from 'react'
 import { createRoot } from 'react-dom/client'
 import { I18nextProvider } from 'react-i18next'
 import { BrowserRouter } from 'react-router-dom'
-import firebase from '@service/firebase'
+import firebase from '@shared/lib/firebase/firebase.ts'
 import i18n from '@shared/config/i18n/index.ts'
 import apolloClient from '@shared/lib/apollo'
 import { baseURL } from '@shared/lib/utils/index.ts'
