+++ apps/web/src/shared/lib/firebase/firebase.ts
